requirements.txt
aiofiles==24.1.0
asyncpg==0.30.0
ccxt==4.4.7
fastapi==0.115.2
gpt4all==2.8.2
httpx==0.27.2
numpy==2.1.1
pandas==2.2.3
plotly==5.24.1
psutil==6.0.0
pydub==0.25.1
pyttsx3==2.91
python-dotenv==1.0.1
sounddevice==0.5.0
speechrecognition==3.10.4
ta==0.11.0
telegram==21.5.0
uvicorn==0.32.0
vosk==0.3.45
wmi==1.5.1
google-api-python-client==2.149.0